Revert "UI: Fixed issue where MSI upgrade would trigger a restart, added error message displaying a clear reason why the update failed when the cause is georestrictions, removed caching within the updater httpClient."
This reverts commit 89056799f441233fe7fa1eff91286c9da64313fe.

+++ ui/src/Update/Update.cs
@@ -32,64 +32,11 @@ namespace FirefoxPrivateNetwork.Update
         private static readonly UpdateHttpClient UpdateHttpClient = new UpdateHttpClient(MaxMsiDownloadSize);
 
         /// <summary>
-        /// Result values for the Run() method.
-        /// </summary>
-        public enum UpdateResult
-        {
-            /// <summary>
-            /// Request has successfully been executed, but there are no new updates.
-            /// </summary>
-            SuccessNoUpdate = 0,
-
-            /// <summary>
-            /// Request has successfully been executed, new updates are available and have been executed.
-            /// </summary>
-            Success,
-
-            /// <summary>
-            /// New updates are available, but the downloaded file's signature is invalid.
-            /// </summary>
-            InvalidSignature,
-
-            /// <summary>
-            /// New updates are available, but we were unable to disconnect before attempting to execute the update process.
-            /// </summary>
-            DisconnectTimeout,
-
-            /// <summary>
-            /// Could not retrieve update information or files due to an HTTP error.
-            /// </summary>
-            HttpError,
-
-            /// <summary>
-            /// New updates are available, signatures are valid, but there was an error when attempting to run the MSI update file.
-            /// </summary>
-            RunFailure,
-
-            /// <summary>
-            /// New updates are available, but we cannot access the MSI update file due to georestrictions.
-            /// </summary>
-            Georestricted,
-
-            /// <summary>
-            /// Unspecified failure.
-            /// </summary>
-            GeneralFailure,
-        }
-
-        private enum UpdateDownloadResult
-        {
-            Success = 0,
-            GeneralFailure,
-            Georestricted,
-        }
-
-        /// <summary>
         /// Starts the update process, downloads the MSI file from Balrog and launches it.
         /// </summary>
         /// <param name="currentVersion">current version of the application.</param>
         /// <returns>Returns an awaitable boolean value indicating whether the update has succeeded or not.</returns>
-        public static async Task<UpdateResult> Run(string currentVersion)
+        public static async Task<bool> Run(string currentVersion)
         {
             var balrogResponse = default(JSONStructures.BalrogResponse);
             try
@@ -99,13 +46,13 @@ namespace FirefoxPrivateNetwork.Update
             catch (HttpRequestException e)
             {
                 ErrorHandling.ErrorHandler.Handle(e, ErrorHandling.LogLevel.Error);
-                return UpdateResult.HttpError;
+                return false;
             }
 
             // Null Balrog response indicates no new update
             if (balrogResponse == null)
             {
-                return UpdateResult.SuccessNoUpdate;
+                return false;
             }
 
             var randomBytes = new byte[32];
@@ -127,7 +74,7 @@ namespace FirefoxPrivateNetwork.Update
                 {
                     ErrorHandling.ErrorHandler.Handle("Unsupported MSI file hash.", ErrorHandling.LogLevel.Error);
                     File.Delete(fileName);
-                    return UpdateResult.InvalidSignature;
+                    return false;
                 }
 
                 // Computes the hash of the file as we write to the file
@@ -135,11 +82,11 @@ namespace FirefoxPrivateNetwork.Update
                 {
                     using (var cryptoStream = new CryptoStream(fileStream, computedHash, CryptoStreamMode.Write))
                     {
-                        var downloadStatus = UpdateDownloadResult.GeneralFailure;
+                        var downloadSuccessful = false;
 
                         try
                         {
-                            downloadStatus = await DownloadAndComputeHashAsync(balrogResponse.MsiUrl, cryptoStream);
+                            downloadSuccessful = await DownloadAndComputeHashAsync(balrogResponse.MsiUrl, cryptoStream);
 
                             // Clean up data streams
                             cryptoStream.FlushFinalBlock();
@@ -149,24 +96,14 @@ namespace FirefoxPrivateNetwork.Update
                         {
                             ErrorHandling.ErrorHandler.Handle(string.Concat("Update MSI file download error: ", e.Message), ErrorHandling.LogLevel.Error);
                             File.Delete(fileName);
-                            return UpdateResult.HttpError;
+                            return false;
                         }
 
-                        switch (downloadStatus)
+                        if (!downloadSuccessful)
                         {
-                            case UpdateDownloadResult.GeneralFailure:
-                                ErrorHandling.ErrorHandler.Handle("Could not download MSI update file.", ErrorHandling.LogLevel.Error);
-                                File.Delete(fileName);
-                                return UpdateResult.GeneralFailure;
-
-                            case UpdateDownloadResult.Georestricted:
-                                ErrorHandling.ErrorHandler.Handle("Could not download MSI update file, georestrictions enforced.", ErrorHandling.LogLevel.Error);
-                                File.Delete(fileName);
-                                return UpdateResult.Georestricted;
-
-                            case UpdateDownloadResult.Success:
-                                ErrorHandling.ErrorHandler.Handle("Successfully downloaded MSI update file.", ErrorHandling.LogLevel.Info);
-                                break;
+                            ErrorHandling.ErrorHandler.Handle("Could not download MSI update file.", ErrorHandling.LogLevel.Error);
+                            File.Delete(fileName);
+                            return false;
                         }
                     }
 
@@ -178,7 +115,7 @@ namespace FirefoxPrivateNetwork.Update
                     {
                         // If the hash comparison fails, delete the file
                         File.Delete(fileName);
-                        return UpdateResult.InvalidSignature;
+                        return false;
                     }
 
                     // Send a disconnect command to the tunnel if the VPN is not already turned off
@@ -199,11 +136,8 @@ namespace FirefoxPrivateNetwork.Update
                     if (waitUntilDisconnected != await Task.WhenAny(waitUntilDisconnected, Task.Delay(DisconnectTimeout)))
                     {
                         File.Delete(fileName);
-                        return UpdateResult.DisconnectTimeout;
+                        return false;
                     }
-
-                    // Shut down broker before updating
-                    Manager.Broker.ShutDown();
 
                     ErrorHandling.ErrorHandler.WriteToLog("Running MSI update...", ErrorHandling.LogLevel.Info);
 
@@ -211,36 +145,28 @@ namespace FirefoxPrivateNetwork.Update
                     if (!msiUpdateLaunch)
                     {
                         File.Delete(fileName);
-                        return UpdateResult.RunFailure;
                     }
 
-                    return UpdateResult.Success;
+                    return msiUpdateLaunch;
                 }
             }
         }
 
-        private static async Task<UpdateDownloadResult> DownloadAndComputeHashAsync(string msiUrl, CryptoStream cryptoStream)
+        private static async Task<bool> DownloadAndComputeHashAsync(string msiUrl, CryptoStream cryptoStream)
         {
             using (var response = await UpdateHttpClient.QueryWithRetryAsync(msiUrl, MaxHttpRetries))
             {
                 if (!response.IsSuccessStatusCode)
                 {
-                    // Georestriction check
-                    if (response.StatusCode == (System.Net.HttpStatusCode)451)
-                    {
-                        ErrorHandling.ErrorHandler.Handle(string.Concat("Error code received while downloading MSI update: ", response.StatusCode), ErrorHandling.LogLevel.Error);
-                        return UpdateDownloadResult.Georestricted;
-                    }
-
                     ErrorHandling.ErrorHandler.Handle(string.Concat("Error code received while downloading MSI update: ", response.StatusCode), ErrorHandling.LogLevel.Error);
-                    return UpdateDownloadResult.GeneralFailure;
+                    return false;
                 }
 
                 // Copy response content to newly created update file
                 await response.Content.CopyToAsync(cryptoStream);
             }
 
-            return UpdateDownloadResult.Success;
+            return true;
         }
 
         private static bool LaunchUpdatedApplication(string fileName)
@@ -253,7 +179,7 @@ namespace FirefoxPrivateNetwork.Update
                     msiProcess.StartInfo.FileName = "msiexec.exe";
 
                     // Runs the MSI installation in basic GUI mode
-                    msiProcess.StartInfo.Arguments = string.Format("/qb!- REBOOT=ReallySuppress /i {0}", Path.GetFileName(fileName));
+                    msiProcess.StartInfo.Arguments = string.Format("/qb!- /i {0}", Path.GetFileName(fileName));
 
                     if (!msiProcess.Start())
                     {
